Adding lshift and rshift
Did a thing where integers can be efficiently left/right shifted. It uses the private field _mp_d for mpz_t, so might not work in later versions. But whatever gets the limbs for that integer (and how many there are) is what you need to pass to mpn_lshift/rshift.